refactor(mobile): single source of truth for folder-tree flatten utils (#1589 finding 2)
Follow-up to PR #1557's review (finding 2). MoveDocumentSheet.flattenTree and
MoveFolderSheet.flattenTreeExcluding were two near-identical depth-first
flattens kept in lock-step by hand (the PR's own test asserted
flattenTreeExcluding(tree, ∅) deep-equals flattenTree(tree)).

Collapse to one shared screens/documents/folderTree.ts holding FlatFolder,
flattenTreeExcluding, collectDescendantIds, and flattenTree (which now delegates
to flattenTreeExcluding with an empty exclusion set — one traversal to maintain).
Both sheets import for internal use and re-export the names, so existing
importers (DocumentFolderOrganization.test.ts, FolderManagement.test.ts) are
unchanged. No behavior change.

(#1589 finding 1 — the "Root (top level)" move silent no-op — is a cross-stack
backend+OpenAPI change tracked on the issue.)
